fix: Rename and restructure code for better readability and organization
This change renames several Cobra command variables from lowercase to
PascalCase, and moves them under a new "RootCmd" variable. It also adds
a new file `commands_test.go` for testing purposes. The overall effect
of this change is to make the code more consistent and easier to
understand by renaming the commands in a more standard manner, and
organizing them under a central RootCmd object. Additionally, a new test
file has been added to help with future testing and maintenance of the
codebase.

Signed-off-by: Alexis von Glasow <vonglasow@gmail.com>