fix: Add unit tests for config package
Implement a suite of tests to verify the functionality and behavior of
the config package in Gaia. The tests cover initializing the
configuration file, using an environment variable for specifying the
configuration file, and setting config values. These tests will help
ensure that the config package works as intended and can be relied upon
in future development.

Signed-off-by: Alexis von Glasow <vonglasow@gmail.com>